Create fuse device to use bindfs
BindFS requires fuse access in order to proceed with the mount. To get
fuse access bindfs needs '/dev/fuse' to exist.
When BindFS is used through vagrant and vagrant-bindfs plugin it tries
to use modprobe to load fuse module, but this is impossible because
loading of kernel modules is not possible inside lxc. Be aware that
fuse kernel module must be loaded on the host machine.